Add splitting signup
